scripts/requirements.txt
# For generate_bot_placeholder.py (used when generating test users and no images are available)
Pillow>=9.0.0
